+++ Tests/scripts/configure_tests.py
@@ -10,9 +10,13 @@ import glob
 import random
 import argparse
 
-from Tests.scripts.constants import *
+SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
+CONTENT_DIR = os.path.abspath(SCRIPT_DIR + '/../..')
+sys.path.append(CONTENT_DIR)
+
+from Tests.scripts.constants import *  # noqa: E402
 from Tests.test_utils import get_yaml, str2bool, get_from_version, get_to_version, \
-    collect_ids, get_script_or_integration_id, run_command, LOG_COLORS, print_error, print_color, print_warning
+    collect_ids, get_script_or_integration_id, run_command, LOG_COLORS, print_error, print_color, print_warning  # noqa: E402
 
 # Search Keyword for the changed file
 NO_TESTS_FORMAT = 'No test( - .*)?'
@@ -143,7 +147,7 @@ def collect_tests(script_ids, playbook_ids, integration_ids, catched_scripts, ca
     caught_missing_test = False
     catched_intergrations = set([])
 
-    test_ids = get_test_ids()
+    test_ids, skipped_tests = get_test_ids()
 
     with open("./Tests/id_set.json", 'r') as conf_file:
         id_set = json.load(conf_file)
@@ -180,7 +184,7 @@ def collect_tests(script_ids, playbook_ids, integration_ids, catched_scripts, ca
                             tests_set.add(test_playbook_id)
                             catched_intergrations.add(integration_id)
 
-        if detected_usage and test_playbook_id not in test_ids:
+        if detected_usage and test_playbook_id not in test_ids and test_playbook_id not in skipped_tests:
             caught_missing_test = True
             print_error("The playbook {} does not appear in the conf.json file, which means no test with it will run."
                         "please update the conf.json file accordingly".format(test_playbook_name))
@@ -201,6 +205,14 @@ def update_missing_sets(catched_intergrations, catched_playbooks, catched_script
 
 
 def get_test_ids(check_nightly_status=False):
+    """Get the test ids from conf.json
+
+    Keyword Arguments:
+        check_nightly_status {bool} -- if we are running nightly test (default: {False})
+
+    Returns:
+        tuple: (test_ids, skipped_tests)
+    """
     test_ids = []
     with open("./Tests/conf.json", 'r') as conf_file:
         conf = json.load(conf_file)
@@ -211,7 +223,7 @@ def get_test_ids(check_nightly_status=False):
             playbook_id = t['playbookID']
             test_ids.append(playbook_id)
 
-    return test_ids
+    return test_ids, list(conf['skipped_tests'].keys())
 
 
 def get_integration_commands(integration_ids, integration_set):
@@ -275,6 +287,11 @@ def update_with_tests_sections(missing_ids, modified_files, test_ids, tests):
 
 
 def collect_changed_ids(integration_ids, playbook_names, script_names, modified_files):
+    tests_set = set([])
+    updated_script_names = set([])
+    updated_playbook_names = set([])
+    catched_scripts, catched_playbooks = set([]), set([])
+
     script_to_version = {}
     playbook_to_version = {}
     integration_to_version = {}
@@ -284,6 +301,12 @@ def collect_changed_ids(integration_ids, playbook_names, script_names, modified_
             name = get_name(file_path)
             script_names.add(name)
             script_to_version[name] = (get_from_version(file_path), get_to_version(file_path))
+
+            package_name = os.path.dirname(file_path)
+            if glob.glob(package_name + "/*_test.py"):
+                catched_scripts.add(name)
+                tests_set.add('Found a unittest for the script {}'.format(package_name))
+
         elif re.match(PLAYBOOK_REGEX, file_path, re.IGNORECASE):
             name = get_name(file_path)
             playbook_names.add(name)
@@ -301,11 +324,6 @@ def collect_changed_ids(integration_ids, playbook_names, script_names, modified_
     script_set = id_set['scripts']
     playbook_set = id_set['playbooks']
     integration_set = id_set['integrations']
-
-    catched_scripts, catched_playbooks = set([]), set([])
-    updated_script_names = set([])
-    updated_playbook_names = set([])
-    tests_set = set([])
 
     for script_id in script_names:
         enrich_for_script_id(script_id, script_to_version[script_id], script_names, script_set, playbook_set,
@@ -384,6 +402,7 @@ def enrich_for_integration_id(integration_id, given_version, integration_command
     for script in script_set:
         script_data = list(script.values())[0]
         script_name = script_data.get('name')
+        script_file_path = script_data.get('file_path')
         script_fromversion = script_data.get('fromversion', '0.0.0')
         script_toversion = script_data.get('toversion', '99.99.99')
         command_to_integration = script_data.get('command_to_integration', {})
@@ -398,6 +417,11 @@ def enrich_for_integration_id(integration_id, given_version, integration_command
                         if tests:
                             catched_scripts.add(script_name)
                             update_test_set(tests, tests_set)
+
+                        package_name = os.path.dirname(script_file_path)
+                        if glob.glob(package_name + "/*_test.py"):
+                            catched_scripts.add(script_name)
+                            tests.add('Found a unittest for the script {}'.format(script_name))
 
                         updated_script_names.add(script_name)
                         new_versions = (script_fromversion, script_toversion)
@@ -433,6 +457,7 @@ def enrich_for_script_id(given_script_id, given_version, script_names, script_se
     for script in script_set:
         script_data = list(script.values())[0]
         script_name = script_data.get('name')
+        script_file_path = script_data.get('file_path')
         script_fromversion = script_data.get('fromversion', '0.0.0')
         script_toversion = script_data.get('toversion', '99.99.99')
         if given_script_id in script_data.get('script_executions', []) and not script_data.get('deprecated') and \
@@ -442,6 +467,11 @@ def enrich_for_script_id(given_script_id, given_version, script_names, script_se
                 if tests:
                     catched_scripts.add(script_name)
                     update_test_set(tests, tests_set)
+
+                package_name = os.path.dirname(script_file_path)
+                if glob.glob(package_name + "/*_test.py"):
+                    catched_scripts.add(script_name)
+                    tests.add('Found a unittest for the script {}'.format(script_name))
 
                 updated_script_names.add(script_name)
                 new_versions = (script_fromversion, script_toversion)
@@ -540,7 +570,7 @@ def get_test_list(files_string, branch_name):
 
     if sample_tests:  # Choosing 3 random tests for infrastructure testing
         print_warning('Collecting sample tests due to: {}'.format(','.join(sample_tests)))
-        test_ids = get_test_ids(check_nightly_status=True)
+        test_ids = get_test_ids(check_nightly_status=True)[0]
         rand = random.Random(files_string + branch_name)
         while len(tests) < 3:
             tests.add(rand.choice(test_ids))