+++ Tests/scripts/wait_until_server_ready.py
@@ -10,7 +10,7 @@ from typing import List, AnyStr
 
 from Tests.test_utils import print_error, print_color, LOG_COLORS
 
-MAX_TRIES = 20
+MAX_TRIES = 30
 SLEEP_TIME = 45