+++ Scripts/StixParser/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-Can now parse single-object STIX2 file.
+You can now parse single-object STIX 2 files.
 
 ## [19.8.2] - 2019-08-22
   - Add indicators: CVE and Registry Key.