+++ Scripts/StixParser/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+-
 
 ## [19.8.2] - 2019-08-22
   - Add indicators: CVE and Registry Key.